Server/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudinary==1.39.0
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dj-database-url==2.1.0
Django==3.2.24
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
fastapi==0.110.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
joblib==1.3.2
jsonschema==4.17.3
kubernetes==29.0.0
llama-index==0.10.16
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.7
llama-index-core==0.10.16.post1
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.7
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.8
llama-index-readers-llama-parse==0.1.3
llama-index-vector-stores-chroma==0.1.5
llama-parse==0.3.6
llamaindex-py-client==0.1.13
Markdown==3.4.4
marshmallow==3.21.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pkgutil_resolve_name==1.3.10
posthog==3.5.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pypdf==4.1.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
sqlparse==0.4.4
starlette==0.36.3
swagger-spec-validator==3.0.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.22.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
whitenoise==6.5.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.15.0
